# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API справочника организаций",
        "description": "Документация REST API справочника организаций, зданий и деятельности",
        "version": "1.0.0"
    },
    "paths": {
        "/api/activities": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Получить список всех видов деятельности",
                "operationId": "53e7a0b67f6097ed90245000764b2e5e",
                "responses": {
                    "200": {
                        "description": "Список видов деятельности",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Activity"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Activities"
                ],
                "summary": "Создать новый вид деятельности",
                "operationId": "8aa12114961e22767ecdad7c24be367e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Программирование"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Вид деятельности создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/activities/{id}": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Получить информацию о виде деятельности",
                "operationId": "6d75cdd5535b9b12358448ec198faba4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о виде деятельности",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Вид деятельности не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Вид деятельности не найден"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Activities"
                ],
                "summary": "Обновить вид деятельности",
                "operationId": "064379988a5fe2c974ae5ab18f0cddfd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Новые технологии"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Вид деятельности обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Activity"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Вид деятельности не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Вид деятельности не найден"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Activities"
                ],
                "summary": "Удалить вид деятельности",
                "operationId": "d822403ad9a85feb951b1073858a072d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Вид деятельности удален"
                    },
                    "400": {
                        "description": "Невозможно удалить вид деятельности с подкатегориями",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Нельзя удалить категорию, у которой есть подкатегории"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Activity": {
                "title": "Activity",
                "description": "Модель деятельности",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Программирование"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-05T10:00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-02-05T10:00:00"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}